Implement inference caching (#79)
We pre-populate a cache containing posterior inferences for the entire
Berry space. The cache is simply an Eigen tensor of vectors.

It takes ~15 seconds to populate the cache. Using the cache, we can run
1 simulation over a 64^4 grid in about 0.8 core seconds.

The cache is not mutated after creation, so it is trivally thread safe.

Most of the code changes is moving functions from SimState to
SimGlobalState (because they are used to compute the global cache).